placed column constructors in different file & export the functions
